Application/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2==2.8.6
pytz==2020.4
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
whitenoise==5.2.0
uWSGI==2.0.19.1